+++ .bak/Makefile.bak
@@ -22,10 +22,14 @@ SRC = \
   src/init_shell/ms_init_env.c \
   src/signal/ms_init_std_signal.c \
   src/signal/ms_stop_std_signal.c \
-  src/ast/create_node.c \
+  src/ast/parse_logical.c \
+  src/ast/parse_pipeline.c \
+  src/ast/parse_subshell.c \
   src/ast/ms_parse_tokens.c \
-  src/ast/parse.c \
+  src/ast/parse_command.c \
+  src/ast/parse_redirection.c \
   src/ast/print_ast.c \
+  src/ast/create_node.c \
   src/ast/utils.c \
   src/global.c \
   src/token/lst_tokens.c \
