+++ src/ast/utils.c
@@ -6,7 +6,7 @@
 /*   By: jeportie <jeportie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/25 14:01:16 by jeportie          #+#    #+#             */
-/*   Updated: 2024/10/25 14:05:21 by jeportie         ###   ########.fr       */
+/*   Updated: 2024/10/29 13:02:42 by jeportie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,4 +23,27 @@ bool	is_logical_op(t_token *current_token)
 	if (current_token->type == TOKEN_AND || current_token->type == TOKEN_OR)
 		return (true);
 	return (false);
+}
+
+bool	is_command_op(t_token *current_token)
+{
+	if (current_token->type == TOKEN_WORD
+		|| current_token->type == TOKEN_EXPAND)
+		return (true);
+	return (false);
+}
+
+bool	is_redir_op(t_token *current_token)
+{
+	return (current_token->type == TOKEN_REDIRECTION);
+}
+
+bool	is_sbs_start(t_token *current_token)
+{
+	return (current_token->type == TOKEN_SUBSHELL_START);
+}
+
+bool	is_sbs_stop(t_token *current_token)
+{
+	return (current_token->type == TOKEN_SUBSHELL_STOP);
 }
